split the rules file into several individual class files
